Create main.py
file_to_binary(): This function reads a file as binary and converts it to a string of 0's and 1's. It then returns the binary string.

binary_to_video(): This function takes a binary string and converts it to a video. Each pixel in the video corresponds to a bit in the binary string. Black pixels represent 1's and white pixels represent 0's.

process_images(): This function takes a list of image frames and processes them to extract a binary string. Each pixel in each frame corresponds to a bit in the binary string. Black pixels represent 1's and white pixels represent 0's.

imageToText(): This function prompts the user to enter the name of a file and converts the file to binary format. It then stores the binary string in a text file.

binaryToFile(): This function takes a binary string as input and converts it back to binary data. It then writes the binary data to an output file.